Prepare 0.2.0 release (API breaking changes)
