+++ Sources/_OpenAPIGeneratorCore/Translator/CommonTranslations/translateAllAnyOneOf.swift
@@ -226,26 +226,6 @@ extension FileTranslator {
             )
         }
 
-        let generateUndocumentedCase = shouldGenerateUndocumentedCaseForEnumsAndOneOfs
-
-        let otherCases: [Declaration]
-        if generateUndocumentedCase {
-            let undocumentedCase: Declaration = .commentable(
-                .doc("Parsed a case that was not defined in the OpenAPI document."),
-                .enumCase(
-                    name: Constants.OneOf.undocumentedCaseName,
-                    kind: .nameWithAssociatedValues([
-                        .init(type: undocumentedType.fullyQualifiedSwiftName)
-                    ])
-                )
-            )
-            otherCases = [
-                undocumentedCase
-            ]
-        } else {
-            otherCases = []
-        }
-
         let encoder = translateOneOfEncoder(caseNames: caseNames)
 
         let comment: Comment? =
@@ -256,7 +236,7 @@ extension FileTranslator {
             accessModifier: config.access,
             name: typeName.shortSwiftName,
             conformances: Constants.ObjectStruct.conformances,
-            members: caseDecls + otherCases + codingKeysDecls + [
+            members: caseDecls + codingKeysDecls + [
                 decoder,
                 encoder,
             ]
