+++ Sources/_OpenAPIGeneratorCore/Translator/CommonTranslations/translateCodable.swift
@@ -339,61 +339,11 @@ extension FileTranslator {
             )
         }
 
-        let generateUndocumentedCase = shouldGenerateUndocumentedCaseForEnumsAndOneOfs
-        let otherExprs: [CodeBlock]
-        if generateUndocumentedCase {
-            otherExprs = [
-                .declaration(
-                    .variable(
-                        kind: .let,
-                        left: "container",
-                        right: .try(
-                            .identifier("decoder")
-                                .dot("singleValueContainer")
-                                .call([])
-                        )
-                    )
-                ),
-                .declaration(
-                    .variable(
-                        kind: .let,
-                        left: "value",
-                        right: .try(
-                            .identifier("container")
-                                .dot("decode")
-                                .call([
-                                    .init(
-                                        label: nil,
-                                        expression:
-                                            .identifier(
-                                                TypeName
-                                                    .valueContainer
-                                                    .fullyQualifiedSwiftName
-                                            )
-                                            .dot("self")
-                                    )
-                                ])
-                        )
-                    )
-                ),
-                .expression(
-                    .assignment(
-                        left: .identifier("self"),
-                        right: .dot(Constants.OneOf.undocumentedCaseName)
-                            .call([
-                                .init(label: nil, expression: .identifier("value"))
-                            ])
-                    )
-                ),
-            ]
-        } else {
-            otherExprs = [
-                .expression(
-                    translateOneOfDecoderThrowOnUnknownExpr()
-                )
-            ]
-        }
-
+        let otherExprs: [CodeBlock] = [
+            .expression(
+                translateOneOfDecoderThrowOnUnknownExpr()
+            )
+        ]
         return decoderInitializer(
             body: (assignExprs).map { .expression($0) } + otherExprs
         )
@@ -450,60 +400,11 @@ extension FileTranslator {
                     ]
                 )
             }
-        let generateUndocumentedCase = shouldGenerateUndocumentedCaseForEnumsAndOneOfs
-        let otherExprs: [CodeBlock]
-        if generateUndocumentedCase {
-            otherExprs = [
-                .declaration(
-                    .variable(
-                        kind: .let,
-                        left: "container",
-                        right: .try(
-                            .identifier("decoder")
-                                .dot("singleValueContainer")
-                                .call([])
-                        )
-                    )
-                ),
-                .declaration(
-                    .variable(
-                        kind: .let,
-                        left: "value",
-                        right: .try(
-                            .identifier("container")
-                                .dot("decode")
-                                .call([
-                                    .init(
-                                        label: nil,
-                                        expression:
-                                            .identifier(
-                                                TypeName
-                                                    .objectContainer
-                                                    .fullyQualifiedSwiftName
-                                            )
-                                            .dot("self")
-                                    )
-                                ])
-                        )
-                    )
-                ),
-                .expression(
-                    .assignment(
-                        left: .identifier("self"),
-                        right: .dot(Constants.OneOf.undocumentedCaseName)
-                            .call([
-                                .init(label: nil, expression: .identifier("value"))
-                            ])
-                    )
-                ),
-            ]
-        } else {
-            otherExprs = [
-                .expression(
-                    translateOneOfDecoderThrowOnUnknownExpr()
-                )
-            ]
-        }
+        let otherExprs: [CodeBlock] = [
+            .expression(
+                translateOneOfDecoderThrowOnUnknownExpr()
+            )
+        ]
         let body: [CodeBlock] = [
             .declaration(.decoderContainerOfKeysVar()),
             .declaration(
@@ -549,27 +450,19 @@ extension FileTranslator {
     func translateOneOfEncoder(
         caseNames: [String]
     ) -> Declaration {
-        let generateUndocumentedCase = shouldGenerateUndocumentedCaseForEnumsAndOneOfs
-        let otherCaseNames: [String]
-        if generateUndocumentedCase {
-            otherCaseNames = [Constants.OneOf.undocumentedCaseName]
-        } else {
-            otherCaseNames = []
-        }
         let switchExpr: Expression = .switch(
             switchedExpression: .identifier("self"),
-            cases: (caseNames + otherCaseNames)
-                .map { caseName in
-                    .init(
-                        kind: .case(.dot(caseName), ["value"]),
-                        body: [
-                            .expression(
-                                .identifier("value")
-                                    .encodeExpr()
-                            )
-                        ]
-                    )
-                }
+            cases: caseNames.map { caseName in
+                .init(
+                    kind: .case(.dot(caseName), ["value"]),
+                    body: [
+                        .expression(
+                            .identifier("value")
+                            .encodeExpr()
+                        )
+                    ]
+                )
+            }
         )
         return encoderFunction(body: [.expression(switchExpr)])
     }
