+++ Tests/OpenAPIGeneratorCoreTests/Translator/Content/Test_ContentSwiftName.swift
@@ -17,28 +17,8 @@ import OpenAPIKit30
 
 final class Test_ContentSwiftName: Test_Core {
 
-    func testExisting() throws {
-        let nameMaker = makeTranslator(featureFlags: []).contentSwiftName
-        let cases: [(String, String)] = [
-            ("application/json", "json"),
-            ("application/x-www-form-urlencoded", "binary"),
-            ("multipart/form-data", "binary"),
-            ("text/plain", "text"),
-            ("*/*", "binary"),
-            ("application/xml", "binary"),
-            ("application/octet-stream", "binary"),
-            ("application/myformat+json", "json"),
-            ("foo/bar", "binary"),
-        ]
-        try _testIdentifiers(cases: cases, nameMaker: nameMaker)
-    }
-
-    func testProposed_multipleContentTypes() throws {
-        let nameMaker = makeTranslator(featureFlags: [
-            .proposal0001,
-            .multipleContentTypes,
-        ])
-        .contentSwiftName
+    func test() throws {
+        let nameMaker = makeTranslator().contentSwiftName
         let cases: [(String, String)] = [
 
             // Short names.
@@ -60,10 +40,6 @@ final class Test_ContentSwiftName: Test_Core {
             ("application/myformat+json", "application_myformat_plus_json"),
             ("foo/bar", "foo_bar"),
         ]
-        try _testIdentifiers(cases: cases, nameMaker: nameMaker)
-    }
-
-    func _testIdentifiers(cases: [(String, String)], nameMaker: (ContentType) -> String) throws {
         for item in cases {
             let contentType = try XCTUnwrap(ContentType(item.0))
             XCTAssertEqual(nameMaker(contentType), item.1, "Case \(item.0) failed")
