+++ Tests/OpenAPIGeneratorReferenceTests/Resources/ReferenceSources/Petstore/Client.swift
@@ -74,7 +74,7 @@ public struct Client: APIProtocol {
                 try converter.setHeaderFieldAsText(
                     in: &request.headerFields,
                     name: "My-Request-UUID",
-                    value: input.headers.My_Request_UUID
+                    value: input.headers.My_hyphen_Request_hyphen_UUID
                 )
                 try converter.setQueryItemAsText(
                     in: &request,
@@ -93,12 +93,12 @@ public struct Client: APIProtocol {
                 switch response.statusCode {
                 case 200:
                     let headers: Operations.listPets.Output.Ok.Headers = .init(
-                        My_Response_UUID: try converter.getRequiredHeaderFieldAsText(
+                        My_hyphen_Response_hyphen_UUID: try converter.getRequiredHeaderFieldAsText(
                             in: response.headerFields,
                             name: "My-Response-UUID",
                             as: Swift.String.self
                         ),
-                        My_Tracing_Header: try converter.getOptionalHeaderFieldAsText(
+                        My_hyphen_Tracing_hyphen_Header: try converter.getOptionalHeaderFieldAsText(
                             in: response.headerFields,
                             name: "My-Tracing-Header",
                             as: Components.Headers.TracingHeader.self
@@ -168,7 +168,7 @@ public struct Client: APIProtocol {
                 try converter.setHeaderFieldAsJSON(
                     in: &request.headerFields,
                     name: "X-Extra-Arguments",
-                    value: input.headers.X_Extra_Arguments
+                    value: input.headers.X_hyphen_Extra_hyphen_Arguments
                 )
                 converter.setAcceptHeader(
                     in: &request.headerFields,
@@ -188,7 +188,7 @@ public struct Client: APIProtocol {
                 switch response.statusCode {
                 case 201:
                     let headers: Operations.createPet.Output.Created.Headers = .init(
-                        X_Extra_Arguments: try converter.getOptionalHeaderFieldAsJSON(
+                        X_hyphen_Extra_hyphen_Arguments: try converter.getOptionalHeaderFieldAsJSON(
                             in: response.headerFields,
                             name: "X-Extra-Arguments",
                             as: Components.Schemas.CodeError.self
@@ -215,7 +215,7 @@ public struct Client: APIProtocol {
                     return .created(.init(headers: headers, body: body))
                 case 400...499:
                     let headers: Components.Responses.ErrorBadRequest.Headers = .init(
-                        X_Reason: try converter.getOptionalHeaderFieldAsText(
+                        X_hyphen_Reason: try converter.getOptionalHeaderFieldAsText(
                             in: response.headerFields,
                             name: "X-Reason",
                             as: Swift.String.self
@@ -288,6 +288,24 @@ public struct Client: APIProtocol {
                             from: response.body,
                             transforming: { value in .json(value) }
                         )
+                    } else if try converter.isMatchingContentType(
+                        received: contentType,
+                        expectedRaw: "text/plain"
+                    ) {
+                        body = try converter.getResponseBodyAsText(
+                            Swift.String.self,
+                            from: response.body,
+                            transforming: { value in .plainText(value) }
+                        )
+                    } else if try converter.isMatchingContentType(
+                        received: contentType,
+                        expectedRaw: "application/octet-stream"
+                    ) {
+                        body = try converter.getResponseBodyAsBinary(
+                            Foundation.Data.self,
+                            from: response.body,
+                            transforming: { value in .binary(value) }
+                        )
                     } else {
                         throw converter.makeUnexpectedContentTypeError(contentType: contentType)
                     }
@@ -318,6 +336,18 @@ public struct Client: APIProtocol {
                         value,
                         headerFields: &request.headerFields,
                         contentType: "application/json; charset=utf-8"
+                    )
+                case let .plainText(value):
+                    request.body = try converter.setRequiredRequestBodyAsText(
+                        value,
+                        headerFields: &request.headerFields,
+                        contentType: "text/plain"
+                    )
+                case let .binary(value):
+                    request.body = try converter.setRequiredRequestBodyAsBinary(
+                        value,
+                        headerFields: &request.headerFields,
+                        contentType: "application/octet-stream"
                     )
                 }
                 return request
@@ -511,7 +541,7 @@ public struct Client: APIProtocol {
                         body = try converter.getResponseBodyAsText(
                             Swift.String.self,
                             from: response.body,
-                            transforming: { value in .text(value) }
+                            transforming: { value in .plainText(value) }
                         )
                     } else {
                         throw converter.makeUnexpectedContentTypeError(contentType: contentType)
