+++ Tests/PetstoreConsumerTests/Common.swift
@@ -15,6 +15,6 @@ import XCTest
 
 extension Operations.listPets.Output {
     static var success: Self {
-        .ok(.init(headers: .init(My_Response_UUID: "abcd"), body: .json([])))
+        .ok(.init(headers: .init(My_hyphen_Response_hyphen_UUID: "abcd"), body: .json([])))
     }
 }
