--- Tests/PetstoreConsumerTestsFFMultipleContentTypes/Generated
@@ -1,1 +1,0 @@
-../OpenAPIGeneratorReferenceTests/Resources/ReferenceSources/Petstore_FF_MultipleContentTypes
